i101: Move location functions to geolocation.js
